% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_regex.R
\name{infer_regex}
\alias{infer_regex}
\title{Infer the regular expression (regex) and other features of a string}
\usage{
infer_regex(x)
}
\arguments{
\item{x}{a string}
}
\value{
a data frame
}
\description{
Infer the regular expression (regex) and other features of a string
}
\examples{
library(purrr)
regex_df <- map_dfr(rownames(mtcars), infer_regex)
all(map2_lgl(regex_df$string, regex_df$regex, ~grepl(.y, .x)))
}
